app/requirements.txt
Django==5.0
uWSGI==2.0.23
gunicorn==23.0.0
whitenoise==6.6.0
django-cors-headers==4.2.0
numpy==2.2.1
Pillow==10.1.0
pandas==2.2.3
scipy==1.14.1
matplotlib==3.10.0
stringdb==0.1.5
boto3==1.33.11
venn==0.1.3 
seaborn==0.13.0
UpSetPlot==0.8.0
scikit-learn==1.6.0
mpmath==1.3.0
openpyxl==3.1.2
setuptools==68.2.2
django-dotenv==1.4.2
combat==0.3.3
MissForest==4.2.2
tqdm==4.67.1
lightgbm==4.3.0
django-ipware==7.0.1
gprofiler-official==1.0.0
pytest==8.3.3
statsmodels==0.14.4